fix installation and move server
