+++ Makefile
@@ -1,7 +1,7 @@
 .PHONY: dev inspect lint format generate-api-client
 
 dev:
-	@python receiptai/client.py receiptai/gmail_mcp.py
+	@python receiptai/server.py receiptai/gmail_mcp.py
 
 inspect:
 	@mcp dev receiptai/gmail_mcp.py
@@ -11,6 +11,11 @@ lint:
 
 format:
 	ruff format .
+
+install-dev:
+	uv venv
+	source .venv/bin/activate
+	uv pip install -e .
 
 generate-api-client:
 	openapi-python-client generate --path ./receiptai/brex_api/brex-expenses.json --output-path ./receiptai/brex_api/expenses --overwrite
