+++ pyproject.toml
@@ -15,6 +15,13 @@ dependencies = [
     "openapi-python-client>=0.24.3",
 ]
 
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["receiptai"]  # list of folders that contain the packages (["."] by default)
+
 [tool.pyright]
 venvPath = "."
 venv = ".venv"
